backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend server for RAG based ChatBot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "huggingface-hub>=0.34.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "uvicorn>=0.35.0",
]
